TST: stats: add tests for mannwhitneyu.  Thanks to Sebastian Pölsterl.
Includes tests for not yet implemented 'greater' and 'two-sided' tests.
The one existing test says the p-value was verified against Matlab/R, but the
`u` return (Mann-Whitney statistic) is not mentioned as being checked.  This is
the one that's being reported as incorrect.

Finally, the old test was named incorrectly and therefore never run.