Version of email text on urgent interventions
I added a rewritten and more specific version of the text I sent to the
mailing list today, as discussed on the hangout.

I also tried to address Nathaniel's point about the importance of
addressing subtle discrimination, particularly against women.  I wasn't
sure what to add to the procedure, but I thought at least it was worth
emphasizing that we are aware of the problem, and that we intend the
procedure to apply.