BUG: use relative import in boost `*_ufunc.pyx` files
The original `setup.py` code was working, because the generated code
went into a newly created directory `_boost/src/` which did not have an
`__init__.py` file. At that point it looked to Cython like it was a set
of standalone `.pyx`/`.pxd` files not part of any package. So an
absolute import succeeded.

After copying those exact same `.pyx`/`.pxd` files to the build
directory, the absolute import failed, because there was a `__init__.py`
file there. Changing it to a relative import then fixed the problem.

Note that the original `setup.py` code won't work anymore, needs to be
re-adjusted.