DEV: meson: add a mesondev.sh build script
This should be integrated into `runtests.py` or a script more
like it, but for now let's keep it separate and simple.

Adding all co-authors here who contributed to some Meson build
improvements on my (Ralf's) fork. Those contributions weren't to this
exact diff, but given that the history of the main Meson PR is rewritten
to be cleaner, this is the commit that we happen to use to add everyone.
Note that the rest of the commits does have proper attribution - the
commit author did do the heavy lifting on a module, tool or CI feature.

Co-authored-by: Gagandeep Singh <gdp.1807@gmail.com>
Co-authored-by: Smit-create <smitlunagariya.mat18@itbhu.ac.in>
Co-authored-by: Matthew Brett <matthew.brett@gmail.com>
Co-authored-by: Dhruv Vats <dhruv25vats@gmail.com>
Co-authored-by: Harsh Mishra <erbeusgriffincasper@gmail.com>