BLD: meson: add `-lm` to add C extensions in a portable manner (#126)
See https://github.com/scipy/scipy/issues/15414, where it turned
out we were missing `-lm` in the distutils based build, and under some
rare circumstances it doesn't get pulled in. It should be missing in the
Meson build as well - we just haven't noticed yet because GCC doesn't
need this typically (while Clang does), or because it's a transitive
dependency that gets pulled in.