WIP: BLD: require Meson 1.3.0, use C17 + GNU extensions if available
This allowed improving the `open_memstream` detection without resorting
to things like:
```
```

Meson 1.3.0 introduces the option of adding a list of C/C++ language
standards, with the higher ones being used if available, and falling
back to the lower one if required.

We've been saying we require C99, but that has mandatory features
(complex types in particular) that MSVC does not support. Those were
made optional in C11, so it's technically more correct to say that
we require C11, even though we don't yet use other C11-specific features
yet. See the Toolchain Roadmap for more details on this.