MAINT: fix some small array API support issues
These surfaced with the more strict reference implementation in
https://github.com/numpy/numpy/pull/25370 and with the addition
of a `__array_namespace__` method to `numpy.ndarray` for NumPy 2.0
(which caused the `is_numpy` utility function to be wrong).
Even if the `is_numpy` issue is also handled by a change in
array-api-compat, this still seems like a logical change to make;
if `x.__name__` is `'numpy'` then `is_numpy(x)` should return True.

[skip cirrus] [skip circle]