MAINT: mark f2py-generated extension modules as safe to run without GIL
This is possible for numpy>=2.1.0, which is the first NumPy version to
(a) support free-threading, and (b) where `f2py` accepts the
`--freethreading-compatible` flag.

Almost all the f2py-generated extension modules have been tested
fairly heavily with `pytest-run-parallel` now and are safe, the last
ones will follow soon.

Note that the GIL will still be re-enabled upon importing pretty much
any SciPy submodule, because we can't yet mark Pythran-generated
modules as safe. This will need work on the Pythran side and a new
Pythran release, so that may miss the 1.15.0 release.
Users can still disable the GIL manually with `python -Xgil=0` or
with the `PYTHON_GIL=0` environment variable.